Got double3 edge case working!!! Manually tested for simple cases. So keep testing around.
Note: Double-3 implementation is not how I traditionally thought of it.
Instead, curr implementation uses offsets AND includes both ends of
lines in the positive and negative directions.
Therefore, this means that encounteredEmpty is ALWAYS true due to the
empty ends being counted.
Miracuously, the count === 4 and encounteredEmpty being T makes the algo
work.

Work on refactoring code so that it doesn't miraculously work - ie to
work as intended. (possibly keep current implementation if cannot).